+++ platform/fdk/src/main/java/fdk/integration/RateWaste.java
@@ -1,13 +1,15 @@
 package fdk.integration;
 
 
+import java.math.BigDecimal;
+
 public class RateWaste {
     public String idRateWaste;
     public String name;
-    public Double coef;
+    public BigDecimal coef;
     public String country;
 
-    public RateWaste(String idRateWaste, String name, Double coef, String country) {
+    public RateWaste(String idRateWaste, String name, BigDecimal coef, String country) {
         this.idRateWaste = idRateWaste;
         this.name = name;
         this.coef = coef;
