+++ platform/fdk/src/main/java/fdk/integration/Ware.java
@@ -1,12 +1,14 @@
 package fdk.integration;
 
 
+import java.math.BigDecimal;
+
 public class Ware {
     public String idWare;
     public String name;
-    public Double price;
+    public BigDecimal price;
 
-    public Ware(String idWare, String name, Double price) {
+    public Ware(String idWare, String name, BigDecimal price) {
         this.idWare = idWare;
         this.name = name;
         this.price = price;
